backend/requirements.txt
# API & Data
httpx>=0.27.0
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
aiosqlite>=0.20.0

# LLM (Kimi API via OpenAI-compatible interface)
openai>=1.12.0

# Scheduling
apscheduler>=3.10.0

# News API
newsapi-python>=0.2.0

# Utils
python-dotenv>=1.0.0
python-dateutil>=2.8.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
